Profiles of RNTuple benchmarks (`iotools/cms, lhcb`) showed ~10-20% of
total runtime is due to allocations in `RPageSource::UnsealPage`.

One typical RNTuple use case is an event loop over a fixed number of
columns. Columns are divided into pages with a fixed size, e.g.

```
Events | Column 0 | Column 1 | Column 2
-------+----------+----------+----------
  1-10   p0(40kB)   p1(80kB)   p2(80kB)
 11-20   p3(40kB)   p4(80kB)   p5(80kB)
  ...      ...        ...        ...
```

Before this change, page allocations were performed every time a page
is flushed from a column:

```
Events | Column 0 | Column 1 | Column 2
-------+----------+----------+----------
         [alloc]    [alloc]    [alloc]
  1-10   p0(40kB)   p1(80kB)   p2(80kB)
        [dealloc]  [dealloc]  [dealloc]
         [alloc]    [alloc]    [alloc]
 11-20   p3(40kB)   p4(80kB)   p5(80kB)
  ...      ...        ...        ...
        [dealloc]  [dealloc]  [dealloc]
```

After this change, allocations are held onto and reused if possible
(up to a certain limit, which for now is hard-coded to be 32 pages).

```
Events | Column 0 | Column 1 | Column 2
-------+----------+----------+----------
         [alloc]    [alloc]    [alloc]
  1-10   p0(40kB)   p1(80kB)   p2(80kB)
 11-20   p3(40kB)   p4(80kB)   p5(80kB)
  ...      ...        ...        ...
        [dealloc]  [dealloc]  [dealloc]
```

On SSD, the following serial results were obtained for the lhcb and cms
benchmarks (average of ten cold runs, 95% confidence interval):

```
cms

Before: 105145us +- 1338us
After:   93669us +- 2077us
%diff:  10.9%

lhcb

Before: 2111977us +-  8371us
After:  1770781us +- 11185us
%diff:  16.2%
```

Todo: 
- [x] Handle non-mappable case in `PopulatePage`
- [x] Investigate multi-threaded case

For multithreading, the current implementation just uses the existing `UnsealPage` mechanism and is unaffected by these changes -- deleted pages are returned to the cache which deallocates them properly. 

I tried adding an internal mutex to `RPageCache` and bumping the number of cached pages to 128 on the benchmark machine but the runtime was about the same as the existing cache-less implementation. Further cache size increases did not show improvements, but as expected only having 32 cached pages for 128 threads was much slower.

```
./lhcb -m (64-cores, 10 cold runs, 95% confidence interval)
no cache: 776188us +- 9918us
cache (128):    780214us +- 11888us
%diff: -0.5%
cache (32):    1369689us +- 33365us
%diff: -76.46%
```