@jalopezg-r00t 2021-07-12
@mxxo, as I said in the ROOT I/O meeting -and if you don't mind-, I can take this :-).  Therefore, I have self-assigned this PR.