@jalopezg-r00t 2022-01-28
Hey @mxxo;  I did not forget about this! ;-)  My plans are to retake this in mid-February or so...