@mxxo 2022-01-28
Hey @jalopezg-r00t, that's great! It will be cool to see how it turns out :)

For what it's worth I seem to remember not seeing the same improvement on spinning disk while benchmarking. But I have no idea of why that would be the case and it might have just been a setup error by me.