@jalopezg-r00t 2022-11-02
> Profiles of RNTuple benchmarks (`iotools/cms, lhcb`) showed ~10-20% of
> total runtime is due to allocations in `RPageSource::UnsealPage`.

@mxxo @jblomer This work will be resumed soon as we suspect that not reusing addresses is affecting RDMA data transfers (which has an impact on the measured throughput in the DAOS backend).  Specifically, we need to confirm that not reusing addresses that appear in the IOVs array in `daos_obj_{fetch,update}()` is related to a higher overhead due to RDMA MR registration.