@jalopezg-git 2023-02-09
Superseded by PR #12279 (closing this draft, in principle).  Many thanks for the initial idea. @mxxo!  :tada: